test: refactor test suite to improve coverage and have consistent script names
