+++ test/scripts/string_integration_test.py
