+++ test/scripts/struct_parser_test.py
@@ -7,7 +7,7 @@
 import struct as pystruct
 import unittest
 
-from libdestruct.c.struct_parser import definition_to_type
+from libdestruct.c.struct_parser import clear_parser_cache, definition_to_type, PARSED_STRUCTS
 from libdestruct import inflater
 
 
@@ -86,6 +86,62 @@ class TypedefTest(unittest.TestCase):
         s = lib.inflate(t, 0)
         self.assertEqual(s.x.value, 0xDEADBEEF)
         self.assertEqual(s.y.value, -42)
+
+
+class AnonymousStructCacheTest(unittest.TestCase):
+    """Anonymous structs must not pollute the parser cache with a None key."""
+
+    def test_none_key_not_in_cache(self):
+        """Anonymous struct should not pollute PARSED_STRUCTS with a None key."""
+        PARSED_STRUCTS.clear()
+
+        definition_to_type("struct { int x; };")
+
+        self.assertNotIn(None, PARSED_STRUCTS)
+
+
+class UnsizedArrayMemberTest(unittest.TestCase):
+    """Parser must handle flexible array members (e.g. int data[]) gracefully."""
+
+    def test_unsized_array_member(self):
+        """Parsing a struct with a flexible array member should not crash."""
+        try:
+            result = definition_to_type("struct test { int count; int data[]; };")
+            self.assertTrue(hasattr(result, '__annotations__'))
+        except (ValueError, TypeError):
+            pass
+        except AttributeError:
+            self.fail("arr_to_type crashed with AttributeError on unsized array - should handle gracefully")
+
+
+class ForwardTypedefTest(unittest.TestCase):
+    """Forward typedef references are a known parser limitation."""
+
+    def setUp(self):
+        clear_parser_cache()
+
+    def tearDown(self):
+        clear_parser_cache()
+
+    def test_chained_typedefs_in_order(self):
+        """Chained typedefs in declaration order must work."""
+        t = definition_to_type("""
+            typedef unsigned int u32;
+            typedef u32 mytype;
+            struct S { mytype x; };
+        """)
+        data = (42).to_bytes(4, "little")
+        s = t.from_bytes(data)
+        self.assertEqual(s.x.value, 42)
+
+    def test_forward_typedef_reference_raises(self):
+        """Forward typedef reference (use before define) must raise a clear error, not crash."""
+        with self.assertRaises((ValueError, TypeError)):
+            definition_to_type("""
+                typedef mytype1 mytype2;
+                typedef unsigned int mytype1;
+                struct S { mytype2 x; };
+            """)
 
 
 if __name__ == "__main__":
